Add TodoForm
When form is submitted, dispatch addTodo Action and clear the form.
TodoForm is an example of a controlled component, which is the
recommended way of handling forms in React. If validation external is
necessary, the form should be changed to dispatch actions when changed
instead of managing the local state in the form